Added losses directory with loss base class
