+++ rl_coach/architectures/tensorflow_components/general_network.py
@@ -27,20 +27,6 @@ from rl_coach.architectures.tensorflow_components.architecture import TensorFlow
 from rl_coach.architectures.tensorflow_components.dnn_model import GeneralModel
 
 
-class GeneralLoss(keras.losses.Loss):
-    def __init__(self, loss_type='MeanSquaredError', **kwargs):
-        self.loss_type = loss_type
-        self.loss_fn = keras.losses.get(self.loss_type)
-        super().__init__(**kwargs)
-
-    def call(self, y_true, y_pred):
-        return self.loss_fn(y_true, y_pred)
-
-    def get_config(self):
-        base_config = super().get_config()
-        return {**base_config, "loss_type": self.loss_type}
-
-
 class GeneralTensorFlowNetwork(TensorFlowArchitecture):
     """
     A generalized version of all possible networks implemented using tensorflow along with the optimizer and loss.
